Create README.md
READ ME
MAR-ALS algorithm gives excellent results: 1) quick convergence (5-50 seconds) under relatively small niteropt (10-170 iterations) and with insignificant residuals; the decrease of LOF (and other error-dependent quantities) is continuous and exponential, 2) the maximal removed systematic error is about 20% of entered P0 values, 3) the algorithm, being iterative, performs automatically a form of sensitivity analysis, 4) the method also self-controls the quality of the whole computational procedure by the help of the selcontrol indicator. The method disposes a plenitude of ways for tracking the evolution of the iteration procedure. There were not observed local optima in the examined problems. Equally, the mathematical model avoids ill-determinacy (ambiguities). The system is protected from equation insufficiency. We proved the method on samples with many minerals (8-10) and with small diversity in measurement data. The results were satisfactory. MAR-ALS algorithm (suitably updated) can be developed to a robust standard-less method for indicating, qualifying, quantifying and diminishing systematic errors in the mineralogical analysis of rocks based on chemical data.